Split Rect methods into separate files
