+++ Runtime/UnityEngine.Rect/Rect.SliceBottom.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: a362f7ab4f1c5904f8f7e5296752c114
+guid: 66ea884cfa3fd7a45a70b39c296df306
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
