+++ Runtime/UnityEngine.Rect/Rect.SliceLeft.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9c2c7c7f4cdcaf94aa33d9766627fc9a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
