+++ Runtime/UnityEngine.Rect/Rect.SliceRight.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a534ece6123931049a83b2a0023f8e7c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
