+++ Runtime/UnityEngine.Rect/Rect.SliceSquare.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 132d69ae30b9f1d4aa0244627178f420
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
