+++ Runtime/UnityEngine.Rect/Rect.SliceTop.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6baaea0db4a1f514e9f857982a82052c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
